Re #1994 (misc): Suppresses implicit conversion warning (on iOS)
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@5626 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/pjsua/pjsua_app_cli.c
@@ -2174,7 +2174,7 @@ static pj_status_t cmd_video_acc_handler(pj_cli_cmd_val *cval)
     case CMD_VIDEO_ACC_CAP_ID:
     case CMD_VIDEO_ACC_REN_ID:
 	{
-	    int dev = pj_strtol(&cval->argv[1]);
+	    int dev = (int)pj_strtol(&cval->argv[1]);
 
 	    if (cmd_id == CMD_VIDEO_ACC_CAP_ID)
 		acc_cfg.vid_cap_dev = dev;
@@ -2203,7 +2203,7 @@ static pj_status_t cmd_enable_vid_rx(pj_cli_cmd_val *cval)
 
     pjsua_call_vid_strm_op_param_default(&param);
 
-    param.med_idx = pj_strtol(&cval->argv[2]);
+    param.med_idx = (int)pj_strtol(&cval->argv[2]);
     if (pjsua_call_get_stream_info(current_call, param.med_idx, &si) ||
 	si.type != PJMEDIA_TYPE_VIDEO)
     {
@@ -2231,7 +2231,7 @@ static pj_status_t cmd_enable_vid_tx(pj_cli_cmd_val *cval)
 
     pjsua_call_vid_strm_op_param_default(&param);
 
-    param.med_idx = pj_strtol(&cval->argv[2]);
+    param.med_idx = (int)pj_strtol(&cval->argv[2]);
 
     status = pjsua_call_set_vid_strm(current_call, op, &param);
     return status;
@@ -2246,7 +2246,8 @@ static pj_status_t cmd_enable_vid_stream(pj_cli_cmd_val *cval,
 
     pjsua_call_vid_strm_op_param_default(&param);
 
-    param.med_idx = cval->argc > 1 ? pj_strtol(&cval->argv[1]) : -1;
+    param.med_idx = cval->argc > 1 ?
+                    (int)pj_strtol(&cval->argv[1]) : -1;
     param.dir = PJMEDIA_DIR_ENCODING_DECODING;
     return pjsua_call_set_vid_strm(current_call, op, &param);
 }
@@ -2256,9 +2257,11 @@ static pj_status_t cmd_set_cap_dev_id(pj_cli_cmd_val *cval)
     pjsua_call_vid_strm_op_param param;
 
     pjsua_call_vid_strm_op_param_default(&param);
-    param.med_idx = cval->argc > 1? pj_strtol(&cval->argv[1]) : -1;
-    param.cap_dev = cval->argc > 2? pj_strtol(&cval->argv[2]) :
-				    PJMEDIA_VID_DEFAULT_CAPTURE_DEV;
+    param.med_idx = cval->argc > 1?
+                    (int)pj_strtol(&cval->argv[1]) : -1;
+    param.cap_dev = cval->argc > 2?
+                    (int)pj_strtol(&cval->argv[2]) :
+                    PJMEDIA_VID_DEFAULT_CAPTURE_DEV;
 
     return pjsua_call_set_vid_strm(current_call,
 				   PJSUA_CALL_VID_STRM_CHANGE_CAP_DEV,
@@ -2279,7 +2282,7 @@ static pj_status_t cmd_vid_device_refresh()
 
 static pj_status_t cmd_vid_device_preview(pj_cli_cmd_val *cval)
 {
-    int dev_id = pj_strtol(&cval->argv[2]);
+    int dev_id = (int)pj_strtol(&cval->argv[2]);
     pj_bool_t on = (pj_ansi_strnicmp(cval->argv[1].ptr, "On", 2) == 0);
 
     if (on) {
@@ -2339,7 +2342,7 @@ static pj_status_t cmd_vid_codec_list()
 
 static pj_status_t cmd_set_vid_codec_prio(pj_cli_cmd_val *cval)
 {
-    int prio = pj_strtol(&cval->argv[2]);
+    int prio = (int)pj_strtol(&cval->argv[2]);
     pj_status_t status;
 
     status = pjsua_vid_codec_set_priority(&cval->argv[1], (pj_uint8_t)prio);
@@ -2355,8 +2358,8 @@ static pj_status_t cmd_set_vid_codec_fps(pj_cli_cmd_val *cval)
     int M, N;
     pj_status_t status;
 
-    M = pj_strtol(&cval->argv[2]);
-    N = pj_strtol(&cval->argv[3]);
+    M = (int)pj_strtol(&cval->argv[2]);
+    N = (int)pj_strtol(&cval->argv[3]);
     status = pjsua_vid_codec_get_param(&cval->argv[1], &cp);
     if (status == PJ_SUCCESS) {
 	cp.enc_fmt.det.vid.fps.num = M;
@@ -2375,8 +2378,8 @@ static pj_status_t cmd_set_vid_codec_bitrate(pj_cli_cmd_val *cval)
     int M, N;
     pj_status_t status;
 
-    M = pj_strtol(&cval->argv[2]);
-    N = pj_strtol(&cval->argv[3]);
+    M = (int)pj_strtol(&cval->argv[2]);
+    N = (int)pj_strtol(&cval->argv[3]);
     status = pjsua_vid_codec_get_param(&cval->argv[1], &cp);
     if (status == PJ_SUCCESS) {
 	cp.enc_fmt.det.vid.avg_bps = M * 1000;
@@ -2395,8 +2398,8 @@ static pj_status_t cmd_set_vid_codec_size(pj_cli_cmd_val *cval)
     int M, N;
     pj_status_t status;
 
-    M = pj_strtol(&cval->argv[2]);
-    N = pj_strtol(&cval->argv[3]);
+    M = (int)pj_strtol(&cval->argv[2]);
+    N = (int)pj_strtol(&cval->argv[3]);
     status = pjsua_vid_codec_get_param(&cval->argv[1], &cp);
     if (status == PJ_SUCCESS) {
 	cp.enc_fmt.det.vid.size.w = M;
@@ -2437,27 +2440,27 @@ static pj_status_t cmd_arrange_vid_win()
 
 static pj_status_t cmd_show_vid_win(pj_cli_cmd_val *cval, pj_bool_t show)
 {
-    pjsua_vid_win_id wid = pj_strtol(&cval->argv[1]);
+    pjsua_vid_win_id wid = (int)pj_strtol(&cval->argv[1]);
     return pjsua_vid_win_set_show(wid, show);
 }
 
 static pj_status_t cmd_move_vid_win(pj_cli_cmd_val *cval)
 {
-    pjsua_vid_win_id wid = pj_strtol(&cval->argv[1]);
+    pjsua_vid_win_id wid = (int)pj_strtol(&cval->argv[1]);
     pjmedia_coord pos;
 
-    pos.x = pj_strtol(&cval->argv[2]);
-    pos.y = pj_strtol(&cval->argv[3]);
+    pos.x = (int)pj_strtol(&cval->argv[2]);
+    pos.y = (int)pj_strtol(&cval->argv[3]);
     return pjsua_vid_win_set_pos(wid, &pos);
 }
 
 static pj_status_t cmd_resize_vid_win(pj_cli_cmd_val *cval)
 {
-    pjsua_vid_win_id wid = pj_strtol(&cval->argv[1]);
+    pjsua_vid_win_id wid = (int)pj_strtol(&cval->argv[1]);
     pjmedia_rect_size size;
 
-    size.w = pj_strtol(&cval->argv[2]);
-    size.h = pj_strtol(&cval->argv[3]);
+    size.w = (int)pj_strtol(&cval->argv[2]);
+    size.h = (int)pj_strtol(&cval->argv[3]);
     return pjsua_vid_win_set_size(wid, &size);
 }
